smarter and faster alarm entering
first off, this changes the first responder behavior so that you can start typing the time right away. more importantly, it also coordinates between the two NSDatePickers much better. when you enter a time, the date is changed to the next day that the specified time is in the future, and when you enter a date, it is rejected if the time in the time picker has already occured on the day you are trying to enter.

for example, say it's 8pm when you open the add alarm window. you set the alarm time to 10:00pm, and since it's currently before 10pm, the date will not change. but then flipping the time to 10:00am will increment the date by one day, so that the date/time reflected by the two pickers is in the future. if you now try to go and flip the date back to today in this example, the date picker will reject it and leave it as tomorrow, since the date/time reflected by the pickers is not in the future.